Extract common logic and types
