+++ core/src/main/java/org/hisp/dhis/android/core/analytics/LegendStrategy.kt
@@ -1,0 +1,7 @@
+package org.hisp.dhis.android.core.analytics
+
+sealed class LegendStrategy {
+    object None : LegendStrategy()
+    object ByDataItem : LegendStrategy()
+    data class Fixed(val legendSetUid: String) : LegendStrategy()
+}
